Final pre-review refactoring and cleanup
