+++ packages/apollo-client/src/__tests__/local-state/general.ts
@@ -928,76 +928,6 @@ describe('Reset/clear store', () => {
       client.resetStore() as Promise<null>;
     },
   );
-
-  xit(
-    "should warn when an initializer created default value can't be found " +
-      "in the cache, and a matching Query resolver can't be found",
-    async done => {
-      const counterQuery = gql`
-        query {
-          counter @client
-        }
-      `;
-
-      const plusMutation = gql`
-        mutation plus {
-          plus @client
-        }
-      `;
-
-      const cache = new InMemoryCache();
-      const initializers = {
-        counter: () => 10,
-      };
-      const client = new ApolloClient({
-        cache,
-        link: ApolloLink.empty(),
-        resolvers: {
-          Query: {
-            counter: () => {},
-          },
-          Mutation: {
-            plus: (_, __, { cache }) => {
-              const { counter } = cache.readQuery({ query: counterQuery });
-              const data = {
-                counter: counter + 1,
-              };
-              cache.writeData({ data });
-              return null;
-            },
-          },
-        },
-        initializers,
-      });
-
-      let realWarn = console.warn;
-      console.warn = message => {
-        unsub.unsubscribe();
-        console.warn = realWarn;
-        done();
-      };
-
-      await client.mutate({ mutation: plusMutation });
-
-      const componentObservable = client.watchQuery({ query: counterQuery });
-
-      let calledOnce = true;
-      const unsub = componentObservable.subscribe({
-        next: data => {
-          try {
-            expect(calledOnce);
-            calledOnce = false;
-          } catch (e) {
-            done.fail(e);
-          }
-        },
-        error: done.fail,
-        complete: done.fail,
-      });
-
-      client.resetStore();
-    },
-  );
 });
 
 describe('Combining client and server state/operations', () => {
